requirements.txt
langchain
langchain_google_genai
langchain_community
python-dotenv
huggingface_hub
faiss-cpu
torch
transformers
sentence-transformers